requirements.txt
fastapi
uvicorn[standard]
pyyaml
psycopg2-binary
pymysql
prometheus_client
pydantic
sqlalchemy
